Begun adding plot visualisation methods.
Plot will be visualised with fake blocks, outlining the entire plot, the will require the plot to at least be 1 block greater than the area of the plot as to not have fake blocks appearing where the player is trying to build.
Will also look into adding the plot visualisation in an area around the player, for this I need to figure out exactly when to update this, as doing it for each movement is not realistic, possibly running it each time the player enter or exits a plot and doing it in a 100 block radius of the player.