Hologram click events.
Determine the click event based on hologram name, the name includes the plot id and the type of hologram. If the type is UNCLAIMED run /claim <plotID> else run /plot info <plotID> the plot info command will open the correct menu already.
Modified /claim to allow plot id as argument, this is optional, else it will look at the plot you're currently standing in, this will allow players to claim plots they are not near.